requirements.txt
certifi==2021.5.30
dj-database-url==0.5.0
django-heroku==0.3.1
gunicorn==20.1.0
whitenoise==5.2.0
wincertstore==0.2
